rename aws-strands/oneagent to opentelemetry, add real oneagent variant
The existing demo used OTLP direct export (not OneAgent), so rename it to
aws-strands/opentelemetry to match what it actually does.

Add a new aws-strands/oneagent demo: FastAPI + Strands SDK + BedrockModel,
instrumented via oneagent.initialize() matching the aws-bedrock/oneagent pattern.

- rename aws-strands/oneagent/ -> aws-strands/opentelemetry/
- rename test to TestAWSStrandsOpenTelemetry, update app path and DQL filter
- add aws-strands/oneagent/: main.py, server.py, Makefile, pyproject.toml, uv.lock
- add test/e2e/aws_strands_oneagent_test.go: startApp + triggerHaiku + auditSpan
- move aws-strands-opentelemetry to otelcol CI matrix, add aws-strands-oneagent to oneagent matrix
