+++ .github/workflows/e2e.yml
@@ -41,6 +41,10 @@ jobs:
             test_run: TestGroqOneAgent
             otel_service_name: groq/oneagent
             ollama_model: tinyllama
+          - name: aws-strands-oneagent
+            app_dir: aws-strands/oneagent
+            test_run: TestAWSStrandsOneAgent
+            otel_service_name: aws-strands/oneagent
 
     steps:
       - name: Checkout
@@ -192,9 +196,9 @@ jobs:
             app_dir: microsoft-agent-framework/opentelemetry
             test_run: TestMicrosoftAgentFrameworkOpenTelemetry
             otel_service_name: microsoft-agent-framework
-          - name: aws-strands-oneagent
-            app_dir: aws-strands/oneagent
-            test_run: TestAWSStrandsOneAgent
+          - name: aws-strands-opentelemetry
+            app_dir: aws-strands/opentelemetry
+            test_run: TestAWSStrandsOpenTelemetry
             otel_service_name: aws-agent-sdk
 
     steps:
