+++ aws-strands/oneagent/pyproject.toml
@@ -1,13 +1,11 @@
 [project]
-name = "aws-agent-sdk"
+name = "aws-strands-oneagent"
 version = "0.1.0"
-description = "Add your description here"
-readme = "README.md"
 requires-python = ">=3.12"
 dependencies = [
-    "fastapi>=0.115.12",
-    "opentelemetry-instrumentation-fastapi>=0.54b1",
-    "opentelemetry-instrumentation-urllib3>=0.54b1",
-    "strands-agents>=0.1.2",
-    "strands-agents-tools>=0.1.1",
+    "strands-agents",
+    "boto3",
+    "oneagent-sdk",
+    "fastapi",
+    "uvicorn[standard]",
 ]
