+++ aws-strands/oneagent/server.py
@@ -1,0 +1,32 @@
+import asyncio
+import oneagent
+from fastapi import FastAPI, HTTPException
+from pydantic import BaseModel
+
+from main import write_haiku
+
+oneagent.initialize()
+
+app = FastAPI(title="Strands Haiku Writer")
+
+
+class HaikuRequest(BaseModel):
+    topic: str
+
+
+class HaikuResponse(BaseModel):
+    topic: str
+    haiku: str
+
+
+@app.get("/health")
+def health():
+    return {"status": "ok"}
+
+
+@app.post("/haiku", response_model=HaikuResponse)
+async def haiku(req: HaikuRequest):
+    if not req.topic.strip():
+        raise HTTPException(status_code=400, detail="topic must not be empty")
+    result = await asyncio.to_thread(write_haiku, req.topic)
+    return HaikuResponse(topic=req.topic, haiku=result)
