+++ aws-strands/opentelemetry/dynatrace.png
