+++ test/e2e/aws_strands_opentelemetry_test.go
@@ -4,16 +4,15 @@ import (
 	"testing"
 )
 
-func TestAWSStrandsOneAgent(t *testing.T) {
-	startApp(t, "aws-strands/oneagent")
-	triggerHaiku(t, true)
+func TestAWSStrandsOpenTelemetry(t *testing.T) {
+	startCLIApp(t, "aws-strands/opentelemetry")
 
 	// gen_ai.bedrock.guardrail.* (AR-017/AR-018/AR-019) are not emitted
 	// because the demo does not configure Bedrock guardrails — expected FAIL in report.
-	auditSpan(t, "aws-strands", "oneagent", BedrockProfile,
+	auditSpan(t, "aws-strands", "opentelemetry", BedrockProfile,
 		`fetch spans, from: now()-10m
-| filter service.name == "aws-strands/oneagent"
-| filter dt.openpipeline.source == "oneagent"
+| filter service.name == "aws-agent-sdk"
+| filter isNotNull(gen_ai.system)
 | filter isNotNull(gen_ai.request.model)
 | limit 1`)
 }
